+++ tests/integration/test_restart.py
@@ -5,13 +5,12 @@
 import logging
 import subprocess
 from pathlib import Path
-from time import sleep
-from typing import Literal
 
 import jubilant
+from cassandra.cluster import ResultSet
+from tenacity import sleep
 
-from integration.helpers.continuous_writes import ContinuousWrites
-from integration.helpers.juju import get_unit_address
+from integration.helpers.cassandra import connect_cql
 
 logger = logging.getLogger(__name__)
 
@@ -30,125 +29,34 @@ def test_deploy(juju: jubilant.Juju, cassandra_charm: Path, app_name: str) -> No
     )
 
 
-def test_graceful_restart_unit(
-    juju: jubilant.Juju, app_name: str, continuous_writes: ContinuousWrites
-) -> None:
-    continuous_writes.start(juju, app_name, replication_factor=3)
-
-    juju.ssh(
-        f"{app_name}/0",
-        "sudo charmed-cassandra.nodetool drain && sudo snap restart charmed-cassandra.daemon",
-    )
-
-    sleep(60)
-    juju.wait(
-        ready=lambda status: jubilant.all_agents_idle(status) and jubilant.all_active(status),
-        delay=20,
-        timeout=600,
-    )
-
-    continuous_writes.stop_and_assert_writes([get_unit_address(juju, app_name, 0)])
-
-
-def test_kill_process(
-    juju: jubilant.Juju, app_name: str, continuous_writes: ContinuousWrites
-) -> None:
-    continuous_writes.start(juju, app_name, replication_factor=3)
-
-    send_control_signal(juju, f"{app_name}/0", "SIGKILL")
-
-    sleep(60)
-    juju.wait(
-        ready=lambda status: jubilant.all_agents_idle(status) and jubilant.all_active(status),
-        delay=20,
-        timeout=600,
-    )
-
-    continuous_writes.stop_and_assert_writes([get_unit_address(juju, app_name, 0)])
-
-
-def test_freeze_process(
-    juju: jubilant.Juju, app_name: str, continuous_writes: ContinuousWrites
-) -> None:
-    continuous_writes.start(
-        juju, app_name, [get_unit_address(juju, app_name, 1)], replication_factor=3
-    )
-
-    continuous_writes.assert_new_writes([get_unit_address(juju, app_name, 0)])
-
-    send_control_signal(juju, f"{app_name}/0", "SIGSTOP")
-
-    sleep(10)
-
-    continuous_writes.assert_new_writes([get_unit_address(juju, app_name, 1)])
-
-    send_control_signal(juju, f"{app_name}/0", "SIGCONT")
-
-    sleep(10)
-    juju.wait(
-        ready=lambda status: jubilant.all_agents_idle(status) and jubilant.all_active(status),
-        delay=20,
-        timeout=600,
-    )
-
-    continuous_writes.stop_and_assert_writes([get_unit_address(juju, app_name, 0)])
-
-
-def test_graceful_restart_cluster(
-    juju: jubilant.Juju, app_name: str, continuous_writes: ContinuousWrites
-) -> None:
-    for i in range(0, 2):
-        juju.ssh(
-            f"{app_name}/{i}",
-            "sudo charmed-cassandra.nodetool drain && sudo snap restart charmed-cassandra.daemon",
+def test_write(juju: jubilant.Juju, app_name: str) -> None:
+    host = juju.status().apps[app_name].units[f"{app_name}/0"].public_address
+    with connect_cql(juju=juju, app_name=app_name, hosts=[host], timeout=300) as session:
+        session.execute(
+            "CREATE KEYSPACE test "
+            "WITH replication = {'class': 'SimpleStrategy', 'replication_factor': 3}",
         )
-
-    sleep(60)
-    juju.wait(
-        ready=lambda status: jubilant.all_agents_idle(status) and jubilant.all_active(status),
-        delay=20,
-        timeout=1800,
-    )
-
-    continuous_writes.start(juju, app_name, [get_unit_address(juju, app_name, 0)])
-    continuous_writes.assert_new_writes([get_unit_address(juju, app_name, 1)])
-    continuous_writes.stop_and_assert_writes([get_unit_address(juju, app_name, 2)])
+        session.set_keyspace("test")
+        session.execute("CREATE TABLE test(message TEXT PRIMARY KEY)")
+        session.execute("INSERT INTO test(message) VALUES ('hello')")
+    sleep(100)
 
 
-def test_lxc_restart_cluster(
-    juju: jubilant.Juju, app_name: str, continuous_writes: ContinuousWrites
-) -> None:
-    continuous_writes.start(juju, app_name, replication_factor=3)
-
+def test_lxc_restart(juju: jubilant.Juju) -> None:
     subprocess.check_call(["lxc", "restart", "--all"])
-
-    sleep(60)
     juju.wait(
         ready=lambda status: jubilant.all_agents_idle(status) and jubilant.all_active(status),
         delay=20,
         timeout=1800,
     )
 
-    continuous_writes.assert_new_writes([get_unit_address(juju, app_name, 0)])
-    continuous_writes.assert_new_writes([get_unit_address(juju, app_name, 1)])
-    continuous_writes.stop_and_assert_writes([get_unit_address(juju, app_name, 2)])
 
-
-def send_control_signal(
-    juju: jubilant.Juju,
-    unit_name: str,
-    signal: Literal["SIGKILL", "SIGSTOP", "SIGCONT"],
-) -> None:
-    subprocess.check_call(
-        [
-            "lxc",
-            "exec",
-            juju.ssh(unit_name, "hostname").strip(),
-            "--",
-            "pkill",
-            "--signal",
-            signal,
-            "-f",
-            "javaagent:/snap/charmed-cassandra/",
-        ]
-    )
+def test_read(juju: jubilant.Juju, app_name: str) -> None:
+    host = juju.status().apps[app_name].units[f"{app_name}/0"].public_address
+    with connect_cql(juju=juju, app_name=app_name, hosts=[host], keyspace="test") as session:
+        res = session.execute("SELECT message FROM test")
+        assert (
+            isinstance(res, ResultSet)
+            and len(res_list := res.all()) == 1
+            and res_list[0].message == "hello"
+        ), "test data written prior aren't found"
